history prideti input redaguojant vietoj form
