+++ frontend/src/components/history/History.js
@@ -1,25 +1,24 @@
-import React, { useState, useEffect } from "react";
-import HistoryTable from "./HistoryTable.js";
+import React, { useState, useEffect, Fragment } from "react";
+import {getAllUsersData, deleteUserIncome, deleteUserExpenses,} from "../../api/libraries/apiLibraries";
+import ReadOnlyRow from "./ReadOnlyRow.js";
+import EditExpenses from "./EditExpenses.js";
+import EditIncome from "./EditIncome.js";
 import "./style/History.css";
-import {
-  getAllUsersData,
-  deleteUserIncome,
-  deleteUserExpenses,
-} from "../../api/libraries/apiLibraries";
 
 function History() {
   const [users, setUsers] = useState([]);
   const [isLoading, setIsLoading] = useState(false);
   const [userID, setUserID] = useState(0);
-  
-  
+  const [editContactId, setEditContactId] = useState(null);
+
+ 
   //   GET method one user data
   function Render() {
     getAllUsersData().then((res) => {
       setUsers(res.data.data.users[0]);
       setUserID(res.data.data.users[0]._id);
       setIsLoading(true);
-  
+      
     });
   }
 
@@ -32,7 +31,7 @@ function History() {
     let { expenses } = users;
 
     let incomeExpenses = [...income, ...expenses];
-
+    
     function sortByDate(a, b) {
       if (a.date_created < b.date_created) {
         return 1;
@@ -58,27 +57,70 @@ function History() {
       Render();
     }
 
+    const handleEditClick = (event, subID, type) => {
+        event.preventDefault();
+        setEditContactId(subID);
+    
+      };
+
+      const handleCancelClick = () => {
+        setEditContactId(null);
+      };
+
     const incomeExpensesSortedByDate = incomeExpenses.sort(sortByDate);
 
     var userIncomeExpenses = incomeExpensesSortedByDate.map((item) => {
       return (
-        <HistoryTable
-          getAllUsersData={getAllUsersData}
-          key={item._id}
-          subID={item._id}
-          date={item.date}
-          category={item.category}
-          amount={item.amount}
-          deleteItem={deleteItem}
-          type={item.type}
-          name={item.name}
-          userID={users._id}
-          Render={Render}
-        />
+        <Fragment>
+        {editContactId === item._id && item.type === "expenses" ? (
+         <EditExpenses 
+            key={item._id}
+            handleCancelClick = {handleCancelClick}
+            subID={item._id}
+            date={item.date}
+            category={item.category}
+            amount={item.amount}
+            type={item.type}
+            name={item.name}
+            userID={users._id}
+            Render={Render}
+            setEditContactId = {setEditContactId}
+          />
+         ) : editContactId === item._id && item.type === "income" ? (
+         <EditIncome 
+            key={item._id}
+            handleCancelClick = {handleCancelClick}
+            subID={item._id}
+            date={item.date}
+            category={item.category}
+            amount={item.amount}
+            type={item.type}
+            name={item.name}
+            userID={users._id}
+            Render={Render}
+            setEditContactId = {setEditContactId}
+            />
+         )
+          : (
+          <ReadOnlyRow 
+            key={item._id}
+            subID={item._id}
+            date={item.date}
+            category={item.category}
+            amount={item.amount}
+            deleteItem={deleteItem}
+            type={item.type}
+            name={item.name}
+            userID={users._id}
+            Render={Render}
+            handleEditClick = {handleEditClick}
+            />
+         )
+        }
+      </Fragment>
       );
     });
   }
-
   return (
     <>
       <div className="History-container">
