--- frontend/src/components/history/style/EditIncome.css
@@ -1,12 +1,0 @@
-.EditHistoryForm-body {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.EditHistoryForm-button {
-  margin: 0.5em 0;
-  display: flex;
-  gap: 10px;
-  justify-content: center;
-}
