+++ frontend/src/components/history/style/History.css
@@ -1,16 +1,41 @@
 .History-container {
   display: flex;
   justify-content: center;
+  max-width: 100%;
+  padding-top: 30px;
 }
 
 .History-body {
-  margin: 2em 0;
-  padding: 2em;
-  box-shadow: 0px 0px 15px 0px rgba(0, 0, 0, 0.15);
-  border: 1px solid black;
-  border-radius: 5px;
-  text-align: center;
-  background-color: #9f9996;
+  border-collapse: collapse;
+  width: 80%;
 }
 
+th,
+td {
+  border: 1px solid #ffffff;
+  text-align: left;
+  padding: 1px;
+  font-size: 23px;
+}
 
+th {
+  background-color: rgb(117, 201, 250);
+}
+
+td {
+  background-color: #ffffff;
+}
+
+.table-income {
+  color: rgb(49, 226, 43);
+}
+
+.table-expense {
+  color: red;
+}
+
+.HistoryTable-button {
+  display: flex;
+  flex-direction: row;
+  gap: 10px;
+}
