--- frontend/src/components/history/style/TableHistory.css
@@ -1,13 +1,0 @@
-.table-income {
-  color: rgb(49, 226, 43);
-}
-
-.table-expense {
-  color: red;
-}
-
-.HistoryTable-button {
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-}
